+++ src/app/modules/blog/blog.service.ts
@@ -1,23 +1,25 @@
-import httpStatus from 'http-status'
-import ApiError from '../../errors/ApiError'
 import { TBlog } from './blog.interface'
 import { Blog } from './blog.model'
 import QueryBuilder from '../../builder/queryBuilder'
 import { searchableFieldsForBlog, selectedFileld } from './blog.constant'
 import { JwtPayload } from 'jsonwebtoken'
+import ApiError from '../../errors/ApiError'
+import httpStatus from 'http-status'
 
+//create blog
 const createBlogIntoDB = async (payload: TBlog) => {
   const createdBlog = await Blog.create(payload)
   const { _id } = createdBlog
   const result = await Blog.findById(_id)
-    .populate('author', 'name email')
+    .populate('author', '-password')
     .select(selectedFileld)
   return result
 }
 
+//find all blogs
 const findAllBlogsFromDB = async (query: Record<string, unknown>) => {
   const blogQuery = new QueryBuilder(
-    Blog.find().populate('author', 'name email').select('title content'),
+    Blog.find().populate('author', '-password').select('title content'),
     query,
   )
     .search(searchableFieldsForBlog)
@@ -36,26 +38,39 @@ const updateBlogIntoDB = async (
   user: JwtPayload,
   payload: TBlog,
 ) => {
+  const isBlogExists = await Blog.findById(id)
+  if (isBlogExists === null || isBlogExists === undefined) {
+    throw new ApiError('Blog not found', httpStatus.NOT_FOUND)
+  }
+
   //check the blog and author who try to update blog
   const isValidBlogAuthor = await Blog.findOne({ _id: id, author: user?.id })
   if (!isValidBlogAuthor) {
-    throw new ApiError('Unauthorized access', httpStatus.UNAUTHORIZED)
+    const error = new Error()
+    error.name = 'AuthorizationError'
+    throw error
   }
 
   //update blog
   const result = await Blog.findByIdAndUpdate(id, payload, { new: true })
-    .populate('author', 'name email')
+    .populate('author', '-password')
     .select(selectedFileld)
   return result
 }
 
 //delete blog
 const deleteBlogFromDB = async (id: string, user: JwtPayload) => {
-  
+  const isBlogExists = await Blog.findById(id)
+  if (isBlogExists === null || isBlogExists === undefined) {
+    throw new ApiError('Blog not found', httpStatus.NOT_FOUND)
+  }
+
   ////check the blog and author who try to delete blog
   const isValidBlogAuthor = await Blog.findOne({ _id: id, author: user?.id })
   if (!isValidBlogAuthor) {
-    throw new ApiError('Unauthorized access', httpStatus.UNAUTHORIZED)
+    const error = new Error()
+    error.name = 'AuthorizationError'
+    throw error
   }
   const result = await Blog.findByIdAndDelete(id)
   return result
